refactor: Use HasBranch with AccessObject

+++ plugins/org.eclipse.osee.ats.core.client/src/org/eclipse/osee/ats/core/client/internal/user/AtsUserServiceClientImpl.java
@@ -133,7 +133,7 @@ public class AtsUserServiceClientImpl extends AbstractAtsUserService implements 
 
    @Override
    public IAtsUser getUserById(long accountId) {
-      return getUserFromOseeUser(UserManager.getUserByArtId((int) accountId));
+      return getUserFromOseeUser(UserManager.getUserByArtId(accountId));
    }
 
    @Override
